Move process_blocksig to indepenedent file
