+++ src/signer_node/message_processor/mod.rs
@@ -6,8 +6,10 @@ mod process_candidateblock;
 mod process_completedblock;
 mod process_nodevss;
 mod process_blockvss;
+mod process_blocksig;
 
 pub use process_candidateblock::process_candidateblock;
 pub use process_completedblock::process_completedblock;
 pub use process_nodevss::process_nodevss;
 pub use process_blockvss::process_blockvss;
+pub use process_blocksig::process_blocksig;
